Re-index HieDb during initialization
Re-index HieDb when halfsp is initialized to ensure that the database exists and
is up-to-date. halfsp would crash upon certain requests if the database did not
exist. Now that the database is created or updated at start-up, this should not
happen. If it does happen anyway (e.g. because the database was deleted while
halfsp was running) then restarting the server will take us back to a good
state.